Refactor order management page and enhance data fetching
- Replaced the main page content with a new OrdersContent component for better organization and separation of concerns.
- Implemented a Suspense fallback for loading states while fetching order data.
- Updated the useOrders hook to improve data fetching logic and handle pagination more effectively.
- Enhanced URL management for order ID and date range filters, improving user experience when applying filters.
- Removed unused code and streamlined the component structure for better maintainability.

These changes collectively enhance the order management functionality and improve the overall user experience.
